requirements.txt
Django == 5.2
Pillow == 10.1.0
Faker == 37.3.0
django-allauth == 65.8.1
django_style == 0.1.1
django-split-settings == 1.3.2
django-docopt-command == 1.1.0
python-decouple == 3.8
django-debug-toolbar == 5.2.0
django-zeal == 2.0.4
django_migration_linter == 5.2.0
# query_counter == 0.3.1
django-stubs == 5.2.0
django-stubs-ext == 5.2.0
django-axes == 7.1.0
django-permissions-policy == 4.26.0
django_widget_tweaks == 1.5.0
argon2-cffi == 23.1.0
pytest == 7.4.3
pytest-cov==6.1.1
pytest-django == 4.6.0
model-bakery == 1.20.4
ruff == 0.11.8
pre-commit==4.2.0
requests == 2.32.3
# For allauth
cryptography == 45.0.2
jwt == 1.3.1
pyjwt == 2.10.1
oauthlib == 3.2.2
requests-oauthlib == 2.0.0
# # For allauth UI
# django-allauth-ui == 1.8.0
# slippers == 0.6.2
